Create a function that allows to interpolate and acces colors.

Create a function that allows to interpolate and acces colors.

color_palette(
  colors_to_interpolate,
  base_colors_set,
  subset = as.vector(NULL),
  reverse = FALSE,
  ...
)

Arguments

colors_to_interpolate

Name of the colors to interpolate. It is the name of a list element as generated by the function generate_colors_to_interpolate().

base_colors_set

A list of colors to interplotate. Output from generate_colors_to_interpolate().

subset

A numeric vector indicating the color's position you want to keep.

reverse

Boolean indicating whether or not to reverse the order of the colors.

...

Other arguments to grDevices::colorRampPalette

Value

A function interpolating colors.

Examples

datafile <- system.file("mycolors.csv", 
                        package = "stadlColorPaletteR")
mycolors <- readr::read_csv(datafile)
#> Rows: 5 Columns: 2
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (2): color_names, hex_code
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.

# combinations with 2 colors
base_cols <- generate_colors_to_interpolate(mycolors, 2)

# interpolate first element of base_cols
mycolors_interpolated <- color_palette(
  colors_to_interpolate = 'pal1',
  base_colors = base_cols
)