Generate all possible combination of n colors from the total number N of colors provided. Will be used later for interpolation.

Generate all possible combination of n colors from the total number N of colors provided. Will be used later for interpolation.

generate_colors_to_interpolate(colors, n)

Arguments

colors

A tibble with color_names and hex_codes as columns.

n

An integer between 1 and number of colors provided.

Value

A list of length C(N,n), each element containing n colors, where C(N,n) denotes the number of n-combination from a set of N elements.

Examples

datafile <- system.file("mycolors.csv", 
                        package = "stadlColorPaletteR")
mycolors <- readr::read_csv(datafile)
#> Rows: 5 Columns: 2
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (2): color_names, hex_code
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.

# combinations with 2 colors
generate_colors_to_interpolate(mycolors, 2)
#> $pal1
#> [1] "#FAFAD2" "#E066FF"
#> 
#> $pal2
#> [1] "#FAFAD2" "#424242"
#> 
#> $pal3
#> [1] "#FAFAD2" "#EE799F"
#> 
#> $pal4
#> [1] "#FAFAD2" "#6B6B6B"
#> 
#> $pal5
#> [1] "#E066FF" "#424242"
#> 
#> $pal6
#> [1] "#E066FF" "#EE799F"
#> 
#> $pal7
#> [1] "#E066FF" "#6B6B6B"
#> 
#> $pal8
#> [1] "#424242" "#EE799F"
#> 
#> $pal9
#> [1] "#424242" "#6B6B6B"
#> 
#> $pal10
#> [1] "#EE799F" "#6B6B6B"
#> 

#combinations with all colors
generate_colors_to_interpolate(mycolors, nrow(mycolors))
#> $pal1
#> [1] "#FAFAD2" "#E066FF" "#424242" "#EE799F" "#6B6B6B"
#>